Add basic directional lighting

+++ src/gfx/api/vulkan/Model.h
@@ -14,14 +14,14 @@ class Model
 public:
     [[nodiscard]] static auto loadObj(const Device& device, const std::filesystem::path& path) -> std::unique_ptr<Model>;
 
-    Model(const Device& device, std::span<const Vertex> vertices, std::span<const uint16_t> indices = {});
+    Model(const Device& device, std::span<const Vertex> vertices, std::span<const uint32_t> indices = {});
 
     auto bind(const vk::CommandBuffer& commandBuffer) const -> void;
     auto draw(const vk::CommandBuffer& commandBuffer) const -> void;
 
 private:
     static auto createVertexBuffer(const Device& device, std::span<const Vertex> vertices) -> std::unique_ptr<Buffer>;
-    static auto createIndexBuffer(const Device& device, std::span<const uint16_t> indices) -> std::unique_ptr<Buffer>;
+    static auto createIndexBuffer(const Device& device, std::span<const uint32_t> indices) -> std::unique_ptr<Buffer>;
 
     const Device& _device;
     const std::unique_ptr<Buffer> _vertexBuffer;
